fixed insert operation id problems

+++ Repositories/IGenericRepository.cs
@@ -6,7 +6,7 @@ namespace AuthProject.Repositories
     {
         Task<IEnumerable<T>> GetAsync<T>(QueryOptions options);
         Task<T> GetByIdAsync<T>(string table, string keyColumn, object id);
-        Task<int> InsertAsync<T>(string table, T entity);
+        Task<long> InsertAsync<T>(string table, T entity);
         Task<int> UpdateAsync<T>(string table, string keyColumn, object id, T entity);
         Task<int> DeleteAsync(string table, string keyColumn, object id);
     }
